rename funcs to app, and fix toolbar resizing
