+++ src/main/python/utils/app.py
@@ -1,6 +1,5 @@
 from fbs_runtime.application_context.PyQt5 import ApplicationContext
-
-importTool = ApplicationContext()
+app = ApplicationContext()
 
 def fileImporter(file):
-    return importTool.get_resource(file)
+    return app.get_resource(file)
