+++ src/main/resources/base/config/items.json
@@ -3,8 +3,8 @@
 		"name": "Ellipse",
 		"icon": "ellipse.png",
 		"object": "QGraphicsEllipseItem",
-		"args": [20, 20, 300, 300]
-	},
+        "args": [20, 20, 300, 300]
+    },
     "airBlownCooler": {
         "name": "Air blown cooler",
         "icon": "AirBlownCooler.png",
@@ -454,12 +454,6 @@
         "icon": "Hydrodesulfurization.png",
         "object": "nodeItem",
         "args": ["hydrodesulfurization"]
-    },
-    "iconsaaa": {
-        "name": "iconsaaa",
-        "icon": "iconsaaa.png",
-        "object": "nodeItem",
-        "args": ["iconsaaa"]
     },
     "impactCrusher": {
         "name": "Impact crusher",
